requirements.txt
numpy==2.0.2
opencv-python==4.10.0.84
onnxruntime==1.19.0
PySide6==6.7.2
pydantic==2.8.2
python-dotenv==1.0.1
psutil==6.0.0

PyYAML==6.0.2
